#505 create allot.py to handle allot functions
