+++ src/f01_road/deal.py
@@ -10,7 +10,8 @@ from src.f00_instrument.dict_toolbox import (
     exists_in_nested_dict,
     del_in_nested_dict,
 )
-from src.f01_road.finance import FundNum, TimeLinePoint, default_fund_pool, allot_scale
+from src.f01_road.allot import allot_scale
+from src.f01_road.finance import FundNum, TimeLinePoint, default_fund_pool
 from src.f01_road.road import (
     AcctName,
     OwnerName,
