+++ src/f01_road/test/test_allot_scale.py
@@ -1,4 +1,4 @@
-from src.f01_road.finance import (
+from src.f01_road.allot import (
     allot_scale,
     _create_allot_dict,
     _allot_missing_scale,
